backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
python-dotenv
pyinstaller
python-dateutil
reportlab
httpx